feat(e2e-tests): Extend sequencer layer with metrics fetching
