+++ apps/e2e-tests/src/process-compose/helpers.ts
@@ -1,13 +1,10 @@
-import { Effect, ParseResult, Schema as S } from 'effect';
+import { ParseResult, Schema as S } from 'effect';
 import { $, execa } from 'execa';
 
 import { logMessage } from '../utils/logs';
 
 import { arrayToObject } from '@blocksense/base-utils/array-iter';
 import { rootDir } from '@blocksense/base-utils/env';
-import { getMetrics } from '../utils/metrics/metrics-fetcher';
-import type { UpdatesToNetwork } from './types';
-import { UpdatesToNetworkMetric } from './types';
 
 const ProcessComposeStatusSchema = S.mutable(
   S.Array(
@@ -66,31 +63,4 @@ export function logTestEnvironmentInfo(
     `${status} test environment ${name}`,
     `${status} time: ${time.toDateString()} ${time.toTimeString()}`,
   );
-}
-
-export function fetchUpdatesToNetworksMetric() {
-  return Effect.gen(function* () {
-    const metrics = yield* getMetrics('http://127.0.0.1:5551/metrics');
-    const updatesToNetworks = metrics.filter(
-      metric => metric.name === 'updates_to_networks',
-    )[0];
-    if (!updatesToNetworks) return null;
-
-    const decoded = S.decodeUnknownSync(UpdatesToNetworkMetric)(
-      updatesToNetworks,
-    );
-    return decoded.metrics.reduce((acc: UpdatesToNetwork, item) => {
-      const networkName = item.labels.Network;
-      const feedId = item.labels.FeedId;
-      const value = item.value;
-
-      if (!acc[networkName]) {
-        acc[networkName] = {};
-      }
-
-      acc[networkName][feedId] = value;
-
-      return acc;
-    }, {} as UpdatesToNetwork);
-  });
 }
